refactor: ReservationTime 패키지 이동 entity->model
