+++ src/main/java/roomescape/dao/ReservationTimeImpl.java
@@ -6,7 +6,7 @@ import java.util.Map;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import roomescape.entity.ReservationTime;
+import roomescape.model.ReservationTime;
 
 @Repository
 public class ReservationTimeImpl implements ReservationTimeDao {
