+++ src/main/java/roomescape/dto/ReservationTimeResponse.java
@@ -2,7 +2,7 @@ package roomescape.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
-import roomescape.entity.ReservationTime;
+import roomescape.model.ReservationTime;
 
 public record ReservationTimeResponse(Long id, @JsonFormat(pattern = "HH:mm", timezone = "Asia/Seoul") LocalTime startAt) {
