+++ src/main/java/roomescape/model/ReservationTime.java
@@ -1,4 +1,4 @@
-package roomescape.entity;
+package roomescape.model;
 
 import java.time.LocalTime;
