+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -3,7 +3,7 @@ package roomescape.service;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.dao.ReservationTimeDao;
-import roomescape.entity.ReservationTime;
+import roomescape.model.ReservationTime;
 
 @Service
 public class ReservationTimeService {
